refactor: Auth, Customer Acceptance Test 분리
